refactor: split MCP investigation tools
