+++ docs/agent-maintainer-hardening-branch-roadmap.md
@@ -1,6 +1,6 @@
 # Agent Maintainer Hardening Branch Roadmap
 
-Current chunk: `refactor/mcp-discovery-tools`
+Current chunk: `refactor/mcp-investigation-tools`
 
 ## Goal
 
@@ -106,6 +106,8 @@ module boundary and keep behavior covered by focused tests.
     tools into their own module while preserving compatibility exports.
   - [x] Move filtered usage discovery, coverage, indexed-content search, and
     token-waste candidate tools into a focused registration module.
+  - [x] Move goal-led investigation, hypothesis testing, action briefs, and
+    strict evidence export into a focused orchestration module.
   - [ ] Split the remaining MCP tool families until the registration module is
     below the configured file-length threshold.
 - Split `src/codex_usage_tracker/reports/api.py` by report family before editing
