+++ src/codex_usage_tracker/cli/mcp_investigations.py
@@ -1,0 +1,197 @@
+"""Goal-led investigation MCP tools."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from codex_usage_tracker.cli.mcp_discovery import _pattern_scan_payload
+from codex_usage_tracker.cli.mcp_runtime import mcp
+from codex_usage_tracker.core.paths import (
+    DEFAULT_ALLOWANCE_PATH,
+    DEFAULT_DB_PATH,
+    DEFAULT_PRICING_PATH,
+    DEFAULT_PROJECTS_PATH,
+)
+from codex_usage_tracker.reports.api import (
+    build_action_brief_report,
+    build_agentic_investigation_report,
+    build_hypothesis_test_report,
+    build_investigation_suggestions_report,
+    build_investigation_walk_report,
+    build_local_evidence_export_report,
+)
+
+
+@mcp.tool()
+def usage_suggest_investigations(
+    goal: str | None = None,
+    since: str | None = None,
+    until: str | None = None,
+    thread: str | None = None,
+    include_archived: bool = False,
+    limit: int | None = 10,
+    privacy_mode: str = "normal",
+) -> dict[str, Any]:
+    """Suggest goal-led usage investigations and next MCP tools."""
+    return build_investigation_suggestions_report(
+        goal=goal,
+        since=since,
+        until=until,
+        thread=thread,
+        include_archived=include_archived,
+        limit=limit,
+        privacy_mode=privacy_mode,
+    ).payload
+
+
+@mcp.tool()
+def usage_investigate(
+    goal: str = "token_waste",
+    since: str | None = None,
+    until: str | None = None,
+    thread: str | None = None,
+    include_archived: bool = False,
+    evidence_limit: int = 5,
+    detail_mode: str = "compact",
+    privacy_mode: str = "normal",
+) -> dict[str, Any]:
+    """Run a goal-led aggregate usage investigation."""
+    return build_agentic_investigation_report(
+        db_path=DEFAULT_DB_PATH,
+        pricing_path=DEFAULT_PRICING_PATH,
+        allowance_path=DEFAULT_ALLOWANCE_PATH,
+        projects_path=DEFAULT_PROJECTS_PATH,
+        goal=goal,
+        since=since,
+        until=until,
+        thread=thread,
+        include_archived=include_archived,
+        evidence_limit=evidence_limit,
+        detail_mode=detail_mode,
+        privacy_mode=privacy_mode,
+    ).payload
+
+
+@mcp.tool()
+def usage_action_brief(
+    goal: str = "token_waste",
+    since: str | None = None,
+    until: str | None = None,
+    thread: str | None = None,
+    include_archived: bool = False,
+    evidence_limit: int = 5,
+    privacy_mode: str = "normal",
+) -> dict[str, Any]:
+    """Return compact aggregate remediation brief with concrete next actions."""
+    return build_action_brief_report(
+        db_path=DEFAULT_DB_PATH,
+        pricing_path=DEFAULT_PRICING_PATH,
+        allowance_path=DEFAULT_ALLOWANCE_PATH,
+        projects_path=DEFAULT_PROJECTS_PATH,
+        goal=goal,
+        since=since,
+        until=until,
+        thread=thread,
+        include_archived=include_archived,
+        evidence_limit=evidence_limit,
+        privacy_mode=privacy_mode,
+    ).payload
+
+
+@mcp.tool()
+def usage_test_hypotheses(
+    question: str,
+    hypotheses: list[str] | None = None,
+    since: str | None = None,
+    until: str | None = None,
+    thread: str | None = None,
+    include_archived: bool = False,
+    evidence_limit: int = 5,
+    privacy_mode: str = "normal",
+) -> dict[str, Any]:
+    """Test usage hypotheses against aggregate/local-index diagnostics."""
+    return build_hypothesis_test_report(
+        db_path=DEFAULT_DB_PATH,
+        pricing_path=DEFAULT_PRICING_PATH,
+        allowance_path=DEFAULT_ALLOWANCE_PATH,
+        projects_path=DEFAULT_PROJECTS_PATH,
+        question=question,
+        hypotheses=hypotheses,
+        since=since,
+        until=until,
+        thread=thread,
+        include_archived=include_archived,
+        evidence_limit=evidence_limit,
+        privacy_mode=privacy_mode,
+    ).payload
+
+
+@mcp.tool()
+def usage_context_bloat_scan(
+    since: str | None = None,
+    until: str | None = None,
+    thread: str | None = None,
+    include_archived: bool = False,
+    min_occurrences: int = 2,
+    limit: int | None = 20,
+    privacy_mode: str = "normal",
+) -> dict[str, Any]:
+    """Find high-token threads with local content/event density."""
+    return _pattern_scan_payload(
+        scan_type="context_bloat",
+        since=since,
+        until=until,
+        thread=thread,
+        include_archived=include_archived,
+        min_occurrences=min_occurrences,
+        limit=limit,
+        privacy_mode=privacy_mode,
+    )
+
+
+@mcp.tool()
+def usage_investigation_walk(
+    question: str,
+    since: str | None = None,
+    until: str | None = None,
+    thread: str | None = None,
+    include_archived: bool = False,
+    min_occurrences: int = 2,
+    evidence_limit: int = 5,
+    privacy_mode: str = "normal",
+) -> dict[str, Any]:
+    """Run a bounded local hypothesis walk over normalized usage evidence."""
+    return build_investigation_walk_report(
+        db_path=DEFAULT_DB_PATH,
+        question=question,
+        since=since,
+        until=until,
+        thread=thread,
+        include_archived=include_archived,
+        min_occurrences=min_occurrences,
+        evidence_limit=evidence_limit,
+        privacy_mode=privacy_mode,
+    ).payload
+
+
+@mcp.tool()
+def usage_local_evidence_export(
+    question: str,
+    since: str | None = None,
+    until: str | None = None,
+    thread: str | None = None,
+    include_archived: bool = False,
+    min_occurrences: int = 2,
+    evidence_limit: int = 5,
+) -> dict[str, Any]:
+    """Return a strict shareable local evidence summary without raw content."""
+    return build_local_evidence_export_report(
+        db_path=DEFAULT_DB_PATH,
+        question=question,
+        since=since,
+        until=until,
+        thread=thread,
+        include_archived=include_archived,
+        min_occurrences=min_occurrences,
+        evidence_limit=evidence_limit,
+    ).payload
